feat(iter11): unified (tabs) + progressive settings + unified inbox (UI layer — PR 2/3)
Merge legacy (client-tabs) and (specialist-tabs) groups into a single
(tabs) group with conditional widgets driven by isSpecialist. Old group
files become Redirect stubs (kept for compat; removed in PR 3).

UI changes:
  - app/(tabs)/_layout.tsx: unified navigator, 5 tabs, "Публичные заявки"
    hidden when user.isSpecialist=false (href: null).
  - app/(tabs)/index.tsx: unified User Dashboard with conditional
    specialist widgets (thread-limit gauge, public-requests feed,
    availability toggle).
  - app/(tabs)/messages.tsx: unified Inbox (three old sources merged,
    labels adapt to isSpecialist).
  - app/(tabs)/public-requests.tsx: moved from (specialist-tabs)/requests.
  - app/(tabs)/requests.tsx: moved from (client-tabs)/requests.
  - app/(tabs)/profile.tsx: redirect to /settings.
  - app/(tabs)/create.tsx, search.tsx: deleted.

Settings:
  - app/settings/index.tsx: single progressive page — Личные данные,
    Специалист (Enable CTA for non-specialists / full editor for
    specialists), Уведомления, Аккаунт. Email-only per SA.
  - settings/client.tsx, settings/specialist.tsx: redirect stubs.

Legacy group stubs (PR3 removes directories):
  - app/(client-tabs)/*, (specialist-tabs)/*: Redirect to /(tabs)/*.

Navigation:
  - SidebarNav: unified USER_BASE_ITEMS + conditional specialist extras.
    Accent driven by isSpecialistUser. Settings path unified.
  - AppHeader: breadcrumbs for /(tabs)/* (legacy aliases kept).
    TABS_ROUTES guard removed — (tabs) is authenticated space.
  - Header.tsx: settings path unified.

Auth/routing:
  - auth/email, auth/otp, index, onboarding/profile: post-auth target
    /(tabs) instead of split role groups. Admin keeps /(admin-tabs).

Legacy role-check replacements:
  - requests/[id]/index, specialists/[id]: isSpecialist flag.

Verification: tsc 0 errors (frontend + backend).

Scope: admin-tabs untouched, backend untouched, legacy dirs kept as
redirect stubs for PR3 cleanup.
