+++ app/(client-tabs)/messages.tsx
@@ -1,356 +1,8 @@
-import { useState, useEffect, useCallback } from "react";
-import {
-  View,
-  Text,
-  FlatList,
-  Pressable,
-  ActivityIndicator,
-  RefreshControl,
-  useWindowDimensions,
-} from "react-native";
-import { SafeAreaView } from "react-native-safe-area-context";
-import { useRouter } from "expo-router";
-import HeaderHome from "@/components/HeaderHome";
-import EmptyState from "@/components/ui/EmptyState";
-import { MessageSquare } from "lucide-react-native";
-import MessengerEmptyPane from "@/components/MessengerEmptyPane";
-import ErrorState from "@/components/ui/ErrorState";
-import Avatar from "@/components/ui/Avatar";
-import InlineChatView from "@/components/InlineChatView";
-import { apiGet } from "@/lib/api";
-import { colors, overlay } from "@/lib/theme";
+import { Redirect } from "expo-router";
 
-interface ThreadItem {
-  id: string;
-  otherUser: {
-    id: string;
-    firstName: string | null;
-    lastName: string | null;
-    avatarUrl: string | null;
-  };
-  request: {
-    id: string;
-    title: string;
-    status: string;
-  };
-  lastMessage: {
-    text: string;
-    createdAt: string;
-  } | null;
-  unreadCount: number;
-  createdAt: string;
-}
-
-function displayName(user: { firstName: string | null; lastName: string | null }): string {
-  const parts = [user.firstName, user.lastName].filter(Boolean);
-  return parts.length > 0 ? parts.join(" ") : "Специалист";
-}
-
-function formatTime(dateStr: string): string {
-  const d = new Date(dateStr);
-  const now = new Date();
-  const isToday = d.toDateString() === now.toDateString();
-
-  if (isToday) {
-    return d.toLocaleTimeString("ru-RU", { hour: "2-digit", minute: "2-digit" });
-  }
-  const diffDays = Math.floor((now.getTime() - d.getTime()) / (1000 * 60 * 60 * 24));
-  if (diffDays < 7) {
-    return d.toLocaleDateString("ru-RU", { weekday: "short" });
-  }
-  return d.toLocaleDateString("ru-RU", { day: "numeric", month: "short" });
-}
-
-function sortThreads(threads: ThreadItem[]): ThreadItem[] {
-  return [...threads].sort((a, b) => {
-    if (a.unreadCount > 0 && b.unreadCount === 0) return -1;
-    if (a.unreadCount === 0 && b.unreadCount > 0) return 1;
-    const aTime = a.lastMessage
-      ? new Date(a.lastMessage.createdAt).getTime()
-      : new Date(a.createdAt).getTime();
-    const bTime = b.lastMessage
-      ? new Date(b.lastMessage.createdAt).getTime()
-      : new Date(b.createdAt).getTime();
-    return bTime - aTime;
-  });
-}
-
-export default function ClientMessages() {
-  const router = useRouter();
-  const { width } = useWindowDimensions();
-  const isDesktop = width >= 640;
-
-  const [threads, setThreads] = useState<ThreadItem[]>([]);
-  const [loading, setLoading] = useState(true);
-  const [refreshing, setRefreshing] = useState(false);
-  const [error, setError] = useState<string | null>(null);
-  const [selectedThreadId, setSelectedThreadId] = useState<string | null>(null);
-
-  const fetchThreads = useCallback(async () => {
-    setError(null);
-    try {
-      const res = await apiGet<{ items: ThreadItem[] }>("/api/threads");
-      setThreads(res.items);
-    } catch (e) {
-      console.error("fetch client threads error:", e);
-      setError("Не удалось загрузить сообщения");
-    } finally {
-      setLoading(false);
-      setRefreshing(false);
-    }
-  }, []);
-
-  useEffect(() => {
-    fetchThreads();
-  }, [fetchThreads]);
-
-  const handleRefresh = useCallback(() => {
-    setRefreshing(true);
-    fetchThreads();
-  }, [fetchThreads]);
-
-  const sorted = sortThreads(threads);
-
-  const renderThread = useCallback(
-    ({ item, onSelect, selected }: { item: ThreadItem; onSelect?: () => void; selected?: boolean }) => {
-      const hasUnread = item.unreadCount > 0;
-      const name = displayName(item.otherUser);
-
-      const handlePress = () => {
-        if (onSelect) {
-          onSelect();
-        } else {
-          router.push(`/threads/${item.id}` as never);
-        }
-      };
-
-      return (
-        <Pressable
-          accessibilityRole="button"
-          accessibilityLabel={`Чат с ${name}`}
-          onPress={handlePress}
-          className="flex-row items-center px-4 border-b border-border active:bg-surface2"
-          style={({ pressed }) => [
-            {
-              backgroundColor: selected
-                ? colors.accentSoft
-                : hasUnread
-                  ? overlay.accent10
-                  : colors.surface,
-              minHeight: 72,
-              borderLeftWidth: hasUnread ? 3 : 0,
-              borderLeftColor: hasUnread ? colors.primary : "transparent",
-              shadowColor: "#000",
-              shadowOffset: { width: 0, height: 1 },
-              shadowOpacity: 0.06,
-              shadowRadius: 3,
-              elevation: 2,
-            },
-            pressed && { opacity: 0.75 },
-          ]}
-        >
-          {/* Avatar with active indicator and unread badge */}
-          <View
-            className="relative mr-3 my-3.5"
-            style={{
-              shadowColor: "#000",
-              shadowOffset: { width: 0, height: 2 },
-              shadowOpacity: 0.1,
-              shadowRadius: 4,
-              elevation: 3,
-            }}
-          >
-            <Avatar
-              name={name}
-              imageUrl={item.otherUser.avatarUrl ?? undefined}
-              size="md"
-            />
-            {/* Online/active indicator */}
-            <View
-              className="absolute bottom-0 right-0 rounded-full bg-success"
-              style={{ width: 12, height: 12, borderWidth: 2, borderColor: "#fff" }}
-            />
-            {hasUnread && (
-              <View
-                className="absolute -top-1 -right-1 rounded-full bg-accent items-center justify-center px-1"
-                style={{ minWidth: 18, height: 18 }}
-              >
-                <Text className="text-xs font-bold text-white">
-                  {item.unreadCount > 99 ? "99+" : item.unreadCount}
-                </Text>
-              </View>
-            )}
-          </View>
-
-          {/* Content */}
-          <View className="flex-1 min-w-0 py-3.5">
-            <View className="flex-row items-center justify-between gap-2">
-              <Text
-                className={`text-base flex-1 flex-shrink ${
-                  hasUnread ? "font-bold text-text-base" : "font-semibold text-text-base"
-                }`}
-                numberOfLines={1}
-              >
-                {name}
-              </Text>
-              {item.lastMessage && (
-                <Text
-                  className={`text-xs flex-shrink-0 ${
-                    hasUnread ? "text-accent font-semibold" : "text-text-dim"
-                  }`}
-                >
-                  {formatTime(item.lastMessage.createdAt)}
-                </Text>
-              )}
-            </View>
-
-            {/* Request title */}
-            <Text className="text-xs text-text-dim mt-0.5" numberOfLines={1}>
-              {item.request.title}
-            </Text>
-
-            {/* Last message preview */}
-            {item.lastMessage ? (
-              <Text
-                className={`text-sm mt-1 ${hasUnread ? "font-semibold text-text-base" : "text-text-mute"}`}
-                numberOfLines={2}
-              >
-                {item.lastMessage.text}
-              </Text>
-            ) : (
-              <Text className="text-sm mt-1 italic text-text-dim" numberOfLines={1}>
-                Нет сообщений
-              </Text>
-            )}
-          </View>
-        </Pressable>
-      );
-    },
-    [router]
-  );
-
-  if (loading) {
-    return (
-      <SafeAreaView className="flex-1 bg-white" edges={["top"]}>
-        <HeaderHome />
-        <View className="flex-1 items-center justify-center">
-          <ActivityIndicator size="large" color={colors.primary} />
-        </View>
-      </SafeAreaView>
-    );
-  }
-
-  if (error) {
-    return (
-      <SafeAreaView className="flex-1 bg-white" edges={["top"]}>
-        <HeaderHome />
-        <View className="flex-1 items-center justify-center">
-          <ErrorState message={error} onRetry={fetchThreads} />
-        </View>
-      </SafeAreaView>
-    );
-  }
-
-  // Desktop: 2-column layout
-  if (isDesktop) {
-    const isWide = width >= 1024;
-    return (
-      <SafeAreaView className="flex-1 bg-white" edges={["top"]}>
-        <HeaderHome />
-        <View className="flex-1 items-center">
-          <View
-            className="flex-1 flex-row w-full"
-            style={{ maxWidth: isWide ? 1200 : "100%", borderWidth: isWide ? 1 : 0, borderColor: colors.border, borderRadius: isWide ? 12 : 0, overflow: "hidden", marginTop: isWide ? 24 : 0, marginBottom: isWide ? 24 : 0 }}
-          >
-          {/* Thread list panel — scrollable left pane */}
-          <View style={{ maxWidth: 320, flex: 1, borderRightWidth: 1, borderRightColor: colors.border }}>
-            <FlatList
-              data={sorted}
-              keyExtractor={(item) => item.id}
-              renderItem={({ item }) =>
-                renderThread({
-                  item,
-                  onSelect: () => setSelectedThreadId(item.id),
-                  selected: item.id === selectedThreadId,
-                })
-              }
-              refreshControl={
-                <RefreshControl
-                  refreshing={refreshing}
-                  onRefresh={handleRefresh}
-                  tintColor={colors.primary}
-                />
-              }
-              contentContainerStyle={{ flexGrow: 1 }}
-              ListEmptyComponent={
-                <EmptyState
-                  icon={MessageSquare}
-                  title="Нет сообщений"
-                  subtitle="Когда специалисты напишут по вашим заявкам, сообщения появятся здесь"
-                  actionLabel="Найти специалиста"
-                  onAction={() => router.push("/specialists" as never)}
-                />
-              }
-            />
-          </View>
-          {/* Chat panel — sticky right pane (no scroll when empty) */}
-          <View className="flex-1">
-            {selectedThreadId ? (
-              <InlineChatView threadId={selectedThreadId} />
-            ) : (
-              <MessengerEmptyPane
-                title="Выберите диалог слева"
-                hint={
-                  sorted.length > 0
-                    ? `У вас ${sorted.length} ${sorted.length === 1 ? "диалог" : sorted.length < 5 ? "диалога" : "диалогов"}. Нажмите любой, чтобы открыть переписку.`
-                    : "Когда специалисты напишут по заявкам, переписки появятся слева."
-                }
-                leftHint="Список диалогов"
-                primary={{
-                  label: "Создать новую заявку",
-                  onPress: () => router.push("/requests/new" as never),
-                  icon: "plus",
-                }}
-                secondary={{
-                  label: "Найти специалиста",
-                  onPress: () => router.push("/specialists" as never),
-                  icon: "sparkles",
-                }}
-              />
-            )}
-          </View>
-          </View>
-        </View>
-      </SafeAreaView>
-    );
-  }
-
-  // Mobile: full-screen list
-  return (
-    <SafeAreaView className="flex-1 bg-white" edges={["top"]}>
-      <HeaderHome />
-      <FlatList
-        data={sorted}
-        keyExtractor={(item) => item.id}
-        renderItem={({ item }) => renderThread({ item })}
-        refreshControl={
-          <RefreshControl
-            refreshing={refreshing}
-            onRefresh={handleRefresh}
-            tintColor={colors.primary}
-          />
-        }
-        ListEmptyComponent={
-          <EmptyState
-            icon={MessageSquare}
-            title="Нет сообщений"
-            subtitle="Когда специалисты напишут по вашим заявкам, сообщения появятся здесь"
-            actionLabel="Найти специалиста"
-            onAction={() => router.push("/specialists" as never)}
-          />
-        }
-        contentContainerStyle={{ flexGrow: 1, paddingBottom: 20 }}
-      />
-    </SafeAreaView>
-  );
+/**
+ * Iter11 PR 2 — legacy redirect. Use /(tabs)/messages instead.
+ */
+export default function ClientMessagesRedirect() {
+  return <Redirect href={"/(tabs)/messages" as never} />;
 }
