+++ app/(tabs)/messages.tsx
@@ -1,114 +1,529 @@
-import { View, Text, Pressable, FlatList, useWindowDimensions } from "react-native";
+import { useState, useEffect, useCallback } from "react";
+import {
+  View,
+  Text,
+  FlatList,
+  Pressable,
+  ActivityIndicator,
+  RefreshControl,
+  useWindowDimensions,
+} from "react-native";
 import { SafeAreaView } from "react-native-safe-area-context";
-import { MessageCircle } from "lucide-react-native";
+import { useRouter } from "expo-router";
+import HeaderHome from "@/components/HeaderHome";
 import EmptyState from "@/components/ui/EmptyState";
-import { AVATAR_COLORS, colors, overlay } from "@/lib/theme";
+import { MessageSquare } from "lucide-react-native";
+import MessengerEmptyPane from "@/components/MessengerEmptyPane";
+import ErrorState from "@/components/ui/ErrorState";
+import Avatar from "@/components/ui/Avatar";
+import InlineChatView from "@/components/InlineChatView";
+import { useAuth } from "@/contexts/AuthContext";
+import { useRequireAuth } from "@/lib/useRequireAuth";
+import { apiGet } from "@/lib/api";
+import { colors, overlay } from "@/lib/theme";
 
-// Tax-domain conversations (specialists ↔ clients on tax matters).
-const CONVERSATIONS = [
-  { id: "1", name: "Алексей К.", avatar: "А", lastMessage: "Подготовил ответ на требование ИФНС, отправил в чат", time: "2 мин назад", unread: true },
-  { id: "2", name: "Мария С.", avatar: "М", lastMessage: "Завтра заседание по камеральной проверке — всё готово", time: "1 ч назад", unread: true },
-  { id: "3", name: "Давид Р.", avatar: "Д", lastMessage: "Стоимость сопровождения выездной — 80 000 ₽", time: "3 ч назад", unread: false },
-  { id: "4", name: "Софья Л.", avatar: "С", lastMessage: "Счёт разблокирован, документы на почте", time: "Вчера", unread: false },
-  { id: "5", name: "Иван П.", avatar: "И", lastMessage: "Можем обсудить оспаривание решения ИФНС?", time: "Вчера", unread: false },
-  { id: "6", name: "Елена Т.", avatar: "Е", lastMessage: "Отчёт по зарубежным счетам готов к подаче", time: "2 дня назад", unread: false },
-];
+/**
+ * Unified Inbox — iter11 UI layer (PR 2/3).
+ *
+ * Replaces (client-tabs)/messages + (specialist-tabs)/threads.
+ * Backend `/api/threads` is role-aware: returns client-side threads when
+ * user.isSpecialist=false, specialist-side threads when true. The screen
+ * shape is the same — only the "other participant" label differs.
+ *
+ * TODO PR3 (post iter11): unify backend to return a single list that
+ * merges both perspectives for specialists who also create their own
+ * tax-help requests, grouped by request (SA requirement).
+ */
 
-function ConversationItem({
-  name,
-  avatar,
-  lastMessage,
-  time,
-  unread,
-  index,
-}: (typeof CONVERSATIONS)[0] & { index: number }) {
-  const avatarColor = AVATAR_COLORS[index % AVATAR_COLORS.length];
-  return (
-    <Pressable
-      accessibilityRole="button"
-      accessibilityLabel={`Чат с ${name}`}
-      className="flex-row items-center active:bg-surface2"
-      style={[
-        { minHeight: 72, paddingVertical: 16, paddingHorizontal: 16, borderBottomWidth: 1, borderBottomColor: colors.border },
-        unread ? { borderLeftWidth: 3, borderLeftColor: colors.primary, backgroundColor: overlay.accent10 } : {},
-      ]}
-    >
-      <View
-        className="w-12 h-12 rounded-full items-center justify-center"
-        style={{
-          backgroundColor: avatarColor,
-          borderWidth: 2,
-          borderColor: '#ffffff',
-          shadowColor: colors.text,
-          shadowOpacity: 0.1,
-          shadowRadius: 4,
-          shadowOffset: { width: 0, height: 1 },
-        }}
-      >
-        <Text className="text-lg font-bold text-white">{avatar}</Text>
-      </View>
-      <View className="flex-1 ml-3">
-        <View className="flex-row justify-between items-center mb-0.5">
-          <Text className={`text-base ${unread ? "font-bold text-accent" : "font-semibold text-text-base"}`}>
-            {name}
-          </Text>
-          <Text className={`text-xs ${unread ? "text-accent font-semibold" : "text-text-dim"}`}>
-            {time}
-          </Text>
-        </View>
-        <View className="flex-row items-center">
-          <Text
-            className={`flex-1 text-sm ${unread ? "text-text-base font-medium" : "text-text-mute"}`}
-            numberOfLines={1}
+interface ThreadItem {
+  id: string;
+  otherUser: {
+    id: string;
+    firstName: string | null;
+    lastName: string | null;
+    avatarUrl: string | null;
+  };
+  request: {
+    id: string;
+    title: string;
+    status: string;
+  };
+  lastMessage: {
+    text: string;
+    createdAt: string;
+  } | null;
+  unreadCount: number;
+  createdAt: string;
+}
+
+function displayName(
+  user: { firstName: string | null; lastName: string | null },
+  fallback: string
+): string {
+  const parts = [user.firstName, user.lastName].filter(Boolean);
+  return parts.length > 0 ? parts.join(" ") : fallback;
+}
+
+function formatTime(dateStr: string): string {
+  const d = new Date(dateStr);
+  const now = new Date();
+  const isToday = d.toDateString() === now.toDateString();
+
+  if (isToday) {
+    return d.toLocaleTimeString("ru-RU", { hour: "2-digit", minute: "2-digit" });
+  }
+  const diffDays = Math.floor((now.getTime() - d.getTime()) / (1000 * 60 * 60 * 24));
+  if (diffDays < 7) {
+    return d.toLocaleDateString("ru-RU", { weekday: "short" });
+  }
+  return d.toLocaleDateString("ru-RU", { day: "numeric", month: "short" });
+}
+
+function sortThreads(threads: ThreadItem[]): ThreadItem[] {
+  return [...threads].sort((a, b) => {
+    if (a.unreadCount > 0 && b.unreadCount === 0) return -1;
+    if (a.unreadCount === 0 && b.unreadCount > 0) return 1;
+    const aTime = a.lastMessage
+      ? new Date(a.lastMessage.createdAt).getTime()
+      : new Date(a.createdAt).getTime();
+    const bTime = b.lastMessage
+      ? new Date(b.lastMessage.createdAt).getTime()
+      : new Date(b.createdAt).getTime();
+    return bTime - aTime;
+  });
+}
+
+export default function UnifiedInbox() {
+  const router = useRouter();
+  const { ready } = useRequireAuth();
+  const { isSpecialistUser } = useAuth();
+  const { width } = useWindowDimensions();
+  const isDesktop = width >= 640;
+
+  const otherPartyFallback = isSpecialistUser ? "Клиент" : "Специалист";
+
+  const [threads, setThreads] = useState<ThreadItem[]>([]);
+  const [loading, setLoading] = useState(true);
+  const [refreshing, setRefreshing] = useState(false);
+  const [error, setError] = useState<string | null>(null);
+  const [selectedThreadId, setSelectedThreadId] = useState<string | null>(null);
+  const [filter, setFilter] = useState<"all" | "unread">("all");
+
+  const fetchThreads = useCallback(async () => {
+    setError(null);
+    try {
+      const res = await apiGet<{ items: ThreadItem[] }>("/api/threads");
+      setThreads(res.items);
+    } catch (e) {
+      console.error("fetch threads error:", e);
+      setError("Не удалось загрузить сообщения");
+    } finally {
+      setLoading(false);
+      setRefreshing(false);
+    }
+  }, []);
+
+  useEffect(() => {
+    if (!ready) return;
+    fetchThreads();
+  }, [ready, fetchThreads]);
+
+  const handleRefresh = useCallback(() => {
+    setRefreshing(true);
+    fetchThreads();
+  }, [fetchThreads]);
+
+  const sorted = sortThreads(threads);
+  const filtered =
+    filter === "unread" ? sorted.filter((t) => t.unreadCount > 0) : sorted;
+  const unreadTotal = threads.reduce((sum, t) => sum + t.unreadCount, 0);
+
+  const renderThread = useCallback(
+    ({
+      item,
+      onSelect,
+      selected,
+    }: {
+      item: ThreadItem;
+      onSelect?: () => void;
+      selected?: boolean;
+    }) => {
+      const hasUnread = item.unreadCount > 0;
+      const name = displayName(item.otherUser, otherPartyFallback);
+
+      const handlePress = () => {
+        if (onSelect) {
+          onSelect();
+        } else {
+          router.push(`/threads/${item.id}` as never);
+        }
+      };
+
+      return (
+        <Pressable
+          accessibilityRole="button"
+          accessibilityLabel={`Чат с ${name}`}
+          onPress={handlePress}
+          className="flex-row items-center px-4 border-b border-border active:bg-surface2"
+          style={({ pressed }) => [
+            {
+              backgroundColor: selected
+                ? colors.accentSoft
+                : hasUnread
+                  ? overlay.accent10
+                  : colors.surface,
+              minHeight: 72,
+              borderLeftWidth: hasUnread ? 3 : 0,
+              borderLeftColor: hasUnread ? colors.primary : "transparent",
+              shadowColor: "#000",
+              shadowOffset: { width: 0, height: 1 },
+              shadowOpacity: 0.06,
+              shadowRadius: 3,
+              elevation: 2,
+            },
+            pressed && { opacity: 0.75 },
+          ]}
+        >
+          <View
+            className="relative mr-3 my-3.5"
+            style={{
+              shadowColor: "#000",
+              shadowOffset: { width: 0, height: 2 },
+              shadowOpacity: 0.1,
+              shadowRadius: 4,
+              elevation: 3,
+            }}
           >
-            {lastMessage}
-          </Text>
-          {unread && (
-            <View className="w-2.5 h-2.5 rounded-full bg-accent ml-2" />
-          )}
+            <Avatar
+              name={name}
+              imageUrl={item.otherUser.avatarUrl ?? undefined}
+              size="md"
+            />
+            <View
+              className="absolute bottom-0 right-0 rounded-full bg-success"
+              style={{
+                width: 12,
+                height: 12,
+                borderWidth: 2,
+                borderColor: "#fff",
+              }}
+            />
+            {hasUnread && (
+              <View
+                className="absolute -top-1 -right-1 rounded-full bg-accent items-center justify-center px-1"
+                style={{ minWidth: 18, height: 18 }}
+              >
+                <Text className="text-xs font-bold text-white">
+                  {item.unreadCount > 99 ? "99+" : item.unreadCount}
+                </Text>
+              </View>
+            )}
+          </View>
+
+          <View className="flex-1 min-w-0 py-3.5">
+            <View className="flex-row items-center justify-between gap-2">
+              <Text
+                className={`text-base flex-1 flex-shrink ${
+                  hasUnread
+                    ? "font-bold text-text-base"
+                    : "font-semibold text-text-base"
+                }`}
+                numberOfLines={1}
+              >
+                {name}
+              </Text>
+              {item.lastMessage && (
+                <Text
+                  className={`text-xs flex-shrink-0 ${
+                    hasUnread ? "text-accent font-semibold" : "text-text-dim"
+                  }`}
+                >
+                  {formatTime(item.lastMessage.createdAt)}
+                </Text>
+              )}
+            </View>
+
+            <Text className="text-xs text-text-dim mt-0.5" numberOfLines={1}>
+              {item.request.title}
+            </Text>
+
+            {item.lastMessage ? (
+              <Text
+                className={`text-sm mt-1 ${hasUnread ? "font-semibold text-text-base" : "text-text-mute"}`}
+                numberOfLines={2}
+              >
+                {item.lastMessage.text}
+              </Text>
+            ) : (
+              <Text
+                className="text-sm mt-1 italic text-text-dim"
+                numberOfLines={1}
+              >
+                Нет сообщений
+              </Text>
+            )}
+          </View>
+        </Pressable>
+      );
+    },
+    [router, otherPartyFallback]
+  );
+
+  if (!ready || loading) {
+    return (
+      <SafeAreaView className="flex-1 bg-white" edges={["top"]}>
+        <HeaderHome />
+        <View className="flex-1 items-center justify-center">
+          <ActivityIndicator size="large" color={colors.primary} />
         </View>
-      </View>
-    </Pressable>
+      </SafeAreaView>
+    );
+  }
+
+  if (error) {
+    return (
+      <SafeAreaView className="flex-1 bg-white" edges={["top"]}>
+        <HeaderHome />
+        <View className="flex-1 items-center justify-center">
+          <ErrorState message={error} onRetry={fetchThreads} />
+        </View>
+      </SafeAreaView>
+    );
+  }
+
+  const emptyState = (
+    <EmptyState
+      icon={MessageSquare}
+      title={isSpecialistUser ? "Пока нет диалогов" : "Нет сообщений"}
+      subtitle={
+        isSpecialistUser
+          ? "Напишите по публичной заявке — переписка появится здесь"
+          : "Когда специалисты напишут по вашим заявкам, сообщения появятся здесь"
+      }
+      actionLabel={isSpecialistUser ? "Смотреть заявки" : "Найти специалиста"}
+      onAction={() =>
+        router.push(
+          isSpecialistUser
+            ? ("/(tabs)/public-requests" as never)
+            : ("/specialists" as never)
+        )
+      }
+    />
+  );
+
+  // Desktop: 2-column layout
+  if (isDesktop) {
+    const isWide = width >= 1024;
+    return (
+      <SafeAreaView className="flex-1 bg-white" edges={["top"]}>
+        <HeaderHome />
+        <View className="flex-1 items-center">
+          <View
+            className="flex-1 flex-row w-full"
+            style={{
+              maxWidth: isWide ? 1200 : "100%",
+              borderWidth: isWide ? 1 : 0,
+              borderColor: colors.border,
+              borderRadius: isWide ? 12 : 0,
+              overflow: "hidden",
+              marginTop: isWide ? 24 : 0,
+              marginBottom: isWide ? 24 : 0,
+            }}
+          >
+            <View
+              style={{
+                maxWidth: 360,
+                flex: 1,
+                borderRightWidth: 1,
+                borderRightColor: colors.border,
+                backgroundColor: colors.surface,
+              }}
+            >
+              <View
+                style={{
+                  paddingHorizontal: 16,
+                  paddingTop: 16,
+                  paddingBottom: 12,
+                  borderBottomWidth: 1,
+                  borderBottomColor: colors.border,
+                }}
+              >
+                <View className="flex-row items-center gap-2 mb-3">
+                  <Text className="text-xl font-bold text-text-base flex-1">
+                    Сообщения
+                  </Text>
+                  {unreadTotal > 0 && (
+                    <View
+                      className="bg-accent rounded-full items-center justify-center"
+                      style={{
+                        minWidth: 24,
+                        height: 24,
+                        paddingHorizontal: 6,
+                      }}
+                    >
+                      <Text className="text-xs font-bold text-white">
+                        {unreadTotal > 99 ? "99+" : unreadTotal}
+                      </Text>
+                    </View>
+                  )}
+                </View>
+                <View className="flex-row gap-2">
+                  <FilterChip
+                    label="Все"
+                    active={filter === "all"}
+                    onPress={() => setFilter("all")}
+                  />
+                  <FilterChip
+                    label="Непрочитанные"
+                    active={filter === "unread"}
+                    onPress={() => setFilter("unread")}
+                  />
+                </View>
+              </View>
+              <FlatList
+                data={filtered}
+                keyExtractor={(item) => item.id}
+                renderItem={({ item }) =>
+                  renderThread({
+                    item,
+                    onSelect: () => setSelectedThreadId(item.id),
+                    selected: item.id === selectedThreadId,
+                  })
+                }
+                refreshControl={
+                  <RefreshControl
+                    refreshing={refreshing}
+                    onRefresh={handleRefresh}
+                    tintColor={colors.primary}
+                  />
+                }
+                contentContainerStyle={{ flexGrow: 1 }}
+                ListEmptyComponent={emptyState}
+              />
+            </View>
+            <View className="flex-1">
+              {selectedThreadId ? (
+                <InlineChatView threadId={selectedThreadId} />
+              ) : (
+                <MessengerEmptyPane
+                  title="Выберите диалог слева"
+                  hint={
+                    sorted.length > 0
+                      ? `У вас ${sorted.length} ${sorted.length === 1 ? "диалог" : sorted.length < 5 ? "диалога" : "диалогов"}. Нажмите любой, чтобы открыть переписку.`
+                      : isSpecialistUser
+                        ? "Напишите по публичной заявке, чтобы начать переписку с клиентом."
+                        : "Когда специалисты напишут по заявкам, переписки появятся слева."
+                  }
+                  leftHint="Список диалогов"
+                  primary={
+                    isSpecialistUser
+                      ? {
+                          label: "Публичные заявки",
+                          onPress: () =>
+                            router.push("/(tabs)/public-requests" as never),
+                          icon: "sparkles",
+                        }
+                      : {
+                          label: "Создать новую заявку",
+                          onPress: () => router.push("/requests/new" as never),
+                          icon: "plus",
+                        }
+                  }
+                  secondary={
+                    isSpecialistUser
+                      ? undefined
+                      : {
+                          label: "Найти специалиста",
+                          onPress: () => router.push("/specialists" as never),
+                          icon: "sparkles",
+                        }
+                  }
+                />
+              )}
+            </View>
+          </View>
+        </View>
+      </SafeAreaView>
+    );
+  }
+
+  // Mobile: full-screen list
+  return (
+    <SafeAreaView className="flex-1 bg-white" edges={["top"]}>
+      <HeaderHome />
+      <FlatList
+        data={filtered}
+        keyExtractor={(item) => item.id}
+        renderItem={({ item }) => renderThread({ item })}
+        refreshControl={
+          <RefreshControl
+            refreshing={refreshing}
+            onRefresh={handleRefresh}
+            tintColor={colors.primary}
+          />
+        }
+        ListHeaderComponent={
+          <View style={{ paddingHorizontal: 16, paddingTop: 12 }}>
+            <View className="flex-row items-center gap-2 mb-3">
+              <Text className="text-2xl font-bold text-text-base flex-1">
+                Сообщения
+              </Text>
+              {unreadTotal > 0 && (
+                <View
+                  className="bg-accent rounded-full items-center justify-center"
+                  style={{ minWidth: 24, height: 24, paddingHorizontal: 6 }}
+                >
+                  <Text className="text-xs font-bold text-white">
+                    {unreadTotal > 99 ? "99+" : unreadTotal}
+                  </Text>
+                </View>
+              )}
+            </View>
+            <View className="flex-row gap-2 mb-3">
+              <FilterChip
+                label="Все"
+                active={filter === "all"}
+                onPress={() => setFilter("all")}
+              />
+              <FilterChip
+                label="Непрочитанные"
+                active={filter === "unread"}
+                onPress={() => setFilter("unread")}
+              />
+            </View>
+          </View>
+        }
+        ListEmptyComponent={emptyState}
+        contentContainerStyle={{ flexGrow: 1, paddingBottom: 20 }}
+      />
+    </SafeAreaView>
   );
 }
 
-export default function MessagesScreen() {
-  const { width } = useWindowDimensions();
-  const isDesktop = width >= 640;
-  const isWideDesktop = width >= 1024;
-  const containerStyle = isDesktop
-    ? { maxWidth: isWideDesktop ? 1200 : 720, width: "100%" as const, alignSelf: "center" as const, paddingHorizontal: isWideDesktop ? 32 : 0 }
-    : undefined;
-
+function FilterChip({
+  label,
+  active,
+  onPress,
+}: {
+  label: string;
+  active: boolean;
+  onPress: () => void;
+}) {
   return (
-    <SafeAreaView className="flex-1 bg-white">
-      <View className="flex-1" style={containerStyle}>
-        <View className="px-4 pt-3 pb-3 border-b border-border bg-white flex-row items-center justify-between">
-          <Text className="text-2xl font-bold text-text-base">Сообщения</Text>
-          <Pressable
-            accessibilityRole="button"
-            accessibilityLabel="Написать новое сообщение"
-            className="bg-accent-soft rounded-full px-3 py-1.5 justify-center items-center"
-            style={{ minHeight: 44, minWidth: 44 }}
-          >
-            <Text className="text-sm font-semibold text-accent">+ Написать</Text>
-          </Pressable>
-        </View>
-
-        <FlatList
-          data={CONVERSATIONS}
-          keyExtractor={(item) => item.id}
-          renderItem={({ item, index }) => <ConversationItem {...item} index={index} />}
-          contentContainerStyle={{ paddingBottom: 32 }}
-          ListEmptyComponent={
-            <EmptyState
-              icon={MessageCircle}
-              title="Нет сообщений"
-              subtitle="Здесь появятся переписки со специалистами и клиентами"
-            />
-          }
-        />
-      </View>
-    </SafeAreaView>
+    <Pressable
+      accessibilityRole="button"
+      accessibilityLabel={label}
+      onPress={onPress}
+      className={`px-4 rounded-full border ${
+        active ? "bg-accent border-accent" : "bg-white border-border"
+      }`}
+      style={{ height: 36, justifyContent: "center" }}
+    >
+      <Text
+        className={`text-sm font-medium ${active ? "text-white" : "text-text-base"}`}
+      >
+        {label}
+      </Text>
+    </Pressable>
   );
 }
